refactor: rename raft_log_v004 to log_store
# Summary

The raft log storage module drops its version suffix and becomes `log_store`.
The crate-root re-export of the external `raft_log` crate is dropped along
with it.

# Details

V004 is the only raft log format this crate implements, so the suffix
distinguished nothing. `RaftLogV004` and friends keep their names for now;
renaming types is a separate change.

The module is named `log_store` rather than the more obvious `raft_log`
because the crate depends on an external crate by that name. A module and an
extern crate sharing a name makes every `use raft_log::...` in the crate
ambiguous under E0659, and there are several. Spelling `::raft_log` at each of
them would be a permanent cost to avoid a one-time naming decision.

`pub use raft_log;` at the crate root existed for a single integration test,
which can name the dependency directly since it is a normal dependency of this
crate.

`raft_log_v004` becomes `log_store` in the public API.
