+++ crates/server/raft-store/src/log_store/util.rs
@@ -14,18 +14,19 @@
 
 use std::io;
 
-use databend_meta_types::raft_types::ErrorSubject;
-use databend_meta_types::raft_types::ErrorVerb;
+use raft_log::api::raft_log_writer::RaftLogWriter;
+use tokio::sync::oneshot;
 
-use crate::raft_log_v004::io_phase::IOPhase;
+use crate::log_store::IODesc;
+use crate::log_store::RaftLogV004;
+use crate::log_store::callback::Callback;
 
-/// Describe the error that occurred during IO operations to RaftLog.
-#[derive(Debug, thiserror::Error)]
-#[error("RaftLogIOError: {verb}-{subject:?}: {ctx}: failed to {phase}; error: {error}")]
-pub struct RaftLogIOError {
-    pub subject: ErrorSubject,
-    pub verb: ErrorVerb,
-    pub phase: IOPhase,
-    pub error: io::Error,
-    pub ctx: String,
+pub async fn blocking_flush(rl: &mut RaftLogV004) -> Result<(), io::Error> {
+    let (tx, rx) = oneshot::channel();
+    let callback = Callback::new_oneshot(tx, IODesc::unknown("blocking_flush(RaftLogV004)"));
+
+    rl.flush(true, Some(callback))?;
+    rx.await
+        .map_err(|_e| io::Error::other("Failed to receive flush completion"))??;
+    Ok(())
 }
