+++ crates/server/raft-store/src/ondisk/ondisk_test.rs
@@ -36,11 +36,11 @@ use tempfile::TempDir;
 use crate::config::RaftConfig;
 use crate::data_version::DataVersion;
 use crate::header::Header;
+use crate::log_store::Cw;
+use crate::log_store::LogStoreMeta;
+use crate::log_store::RaftLogV004;
 use crate::ondisk::OnDisk;
 use crate::ondisk::TREE_HEADER;
-use crate::raft_log_v004::Cw;
-use crate::raft_log_v004::LogStoreMeta;
-use crate::raft_log_v004::RaftLogV004;
 use crate::sled_compat::LogMetaKey;
 use crate::sled_compat::LogMetaValue;
 use crate::sled_compat::RaftStateKey;
