+++ crates/server/service/src/metrics/meta_metrics.rs
@@ -35,7 +35,7 @@ use prometheus_client::encoding::text::encode as prometheus_encode;
 pub mod server_metrics {
     use std::sync::LazyLock;
 
-    use databend_meta_raft_store::raft_log_v004::RaftLogStat;
+    use databend_meta_raft_store::log_store::RaftLogStat;
     use databend_meta_types::raft_types::NodeId;
     use prometheus_client::metrics::counter::Counter;
     use prometheus_client::metrics::family::Family;
